исправил ошибки

+++ src/main/java/ru/yandex/practicum/filmorate/model/Film.java
@@ -8,10 +8,10 @@ import java.time.LocalDate;
 @Data
 @AllArgsConstructor
 public class Film {
-    private int id;
+    private long id;
     private String name;
     private String description;
     private LocalDate releaseDate;
-    private int duration;
+    private long duration;
 
 }
